Testit inttimuotoon

+++ src/main/java/model/classes/Order.java
@@ -5,7 +5,7 @@ import java.util.ArrayList;
 
 public class Order implements IOrder {
     private ArrayList<Product> productList = new ArrayList<>();
-    private float totalPrice = 0f;
+    private int totalPrice = 0;
 
     @Override
     public ArrayList<Product> getProductList() { return this.productList; }
